fix(server): use BIGINT for monitor minute columns on Postgres
checkin_margin and max_runtime are read as Option<i64> but were declared
INTEGER (INT4) in the Postgres migration, which panics on decode under real
Postgres (INT8 vs INT4) while passing on SQLite's loose typing. Declare them
BIGINT to match.

Also switch the monitor test verification queries from '?' to '$N'
placeholders so the suite runs green under both SQLite and Postgres
(cargo test --features postgres), catching this class of dialect bug.

+++ apps/server/migrations/postgres/20260629000000_create_monitors.up.sql
@@ -13,8 +13,8 @@ CREATE TABLE monitors (
     schedule_type        VARCHAR(16),   -- 'crontab' | 'interval'
     schedule_value       TEXT,          -- crontab string, or interval count as text
     schedule_unit        VARCHAR(8),    -- interval unit: year/month/week/day/hour/minute
-    checkin_margin       INTEGER,       -- minutes after expected time before "missed"
-    max_runtime          INTEGER,       -- minutes an in_progress run may last before "timeout"
+    checkin_margin       BIGINT,        -- minutes after expected time before "missed"
+    max_runtime          BIGINT,        -- minutes an in_progress run may last before "timeout"
     timezone             VARCHAR(64),
     owner                VARCHAR(255),
